separated authorization and work with playlists. Added authorization for every request.
