+++ src/main/java/com/example/spotikservice/controller/CallbackController.java
@@ -1,5 +1,6 @@
 package com.example.spotikservice.controller;
 
+import com.example.spotikservice.service.AuthService;
 import com.example.spotikservice.service.SpotifyService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -16,35 +17,36 @@ import java.util.Map;
 @RequestMapping
 @RequiredArgsConstructor
 public class CallbackController {
-    private final SpotifyService service;
+    private final SpotifyService spotifyService;
+    private final AuthService authService;
 
     @GetMapping
     public int setAccessToken(@RequestParam String code, HttpServletResponse response) {
-        return service.setAccessToken(code, response);
+        return authService.setAccessToken(code, response);
     }
 
     @GetMapping("playlists")
     public PlaylistSimplified[] getPlaylists(HttpServletRequest request) {
-        return service.getPlaylists(request);
+        return spotifyService.getPlaylists(request);
     }
 
     @GetMapping("songs")
-    public Map<String, List<AlbumSimplified>> getFollowedArtists() {
-        return service.getLastReleasesFromSubscribedArtists();
+    public Map<String, List<AlbumSimplified>> getFollowedArtists(HttpServletRequest request) {
+        return spotifyService.getLastReleasesFromSubscribedArtists(request);
     }
 
     @GetMapping("show-ru")
-    public List<PlaylistTrack> checkIfThereAreRussianTracksAdded(@RequestParam String id) {
-        return service.getRussianTracks(id);
+    public List<PlaylistTrack> checkIfThereAreRussianTracksAdded(HttpServletRequest request, @RequestParam String id) {
+        return spotifyService.getRussianTracks(id, request);
     }
 
     @DeleteMapping("remove-all-ru-tracks")
-    public void removeAllRuTracks(@RequestParam String id) {
-        service.removeAllRussianTracksFromPlaylist(id);
+    public void removeAllRuTracks(HttpServletRequest request, @RequestParam String id) {
+        spotifyService.removeAllRussianTracksFromPlaylist(id, request);
     }
 
     @DeleteMapping("remove-track-from-playlist")
-    public void removeTrackFromPlaylist(@RequestParam(name = "playlist-id") String playlistId, @RequestParam(name = "track-id") String trackId) {
-        service.removeTrackFromPlaylist(playlistId, trackId);
+    public void removeTrackFromPlaylist(HttpServletRequest request, @RequestParam(name = "playlist-id") String playlistId, @RequestParam(name = "track-id") String trackId) {
+        spotifyService.removeTrackFromPlaylist(request, playlistId, trackId);
     }
 }
