+++ src/main/java/com/example/spotikservice/service/AuthService.java
@@ -1,5 +1,12 @@
 package com.example.spotikservice.service;
 
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+
 public interface AuthService {
     String authorize();
+
+    int setAccessToken(String code, HttpServletResponse response);
+
+    void setAccessToken(HttpServletRequest request);
 }
