+++ src/main/java/com/example/spotikservice/service/SpotifyService.java
@@ -1,7 +1,6 @@
 package com.example.spotikservice.service;
 
 import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
 import se.michaelthelin.spotify.model_objects.specification.AlbumSimplified;
 import se.michaelthelin.spotify.model_objects.specification.PlaylistSimplified;
 import se.michaelthelin.spotify.model_objects.specification.PlaylistTrack;
@@ -12,13 +11,11 @@ import java.util.Map;
 public interface SpotifyService {
     PlaylistSimplified[] getPlaylists(HttpServletRequest request);
 
-    int setAccessToken(String code, HttpServletResponse response);
+    Map<String, List<AlbumSimplified>> getLastReleasesFromSubscribedArtists(HttpServletRequest request);
 
-    Map<String, List<AlbumSimplified>> getLastReleasesFromSubscribedArtists();
+    List<PlaylistTrack> getRussianTracks(String id, HttpServletRequest request);
 
-    List<PlaylistTrack> getRussianTracks(String id);
+    void removeAllRussianTracksFromPlaylist(String playlistId, HttpServletRequest request);
 
-    void removeAllRussianTracksFromPlaylist(String playlistId);
-
-    void removeTrackFromPlaylist(String playlistId, String trackId);
+    void removeTrackFromPlaylist(HttpServletRequest request, String playlistId, String trackId);
 }
