+++ src/main/java/com/example/spotikservice/service/impl/SpotifyServiceImpl.java
@@ -2,31 +2,21 @@ package com.example.spotikservice.service.impl;
 
 import com.example.spotikservice.constants.CacheConstants;
 import com.example.spotikservice.dao.SpotifyArtistDao;
-import com.example.spotikservice.dao.UserDao;
 import com.example.spotikservice.entities.SpotifyArtist;
-import com.example.spotikservice.entities.User;
-import com.example.spotikservice.exception.UserUnauthorizedException;
+import com.example.spotikservice.service.AuthService;
 import com.example.spotikservice.service.SpotifyService;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
-import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
-import org.apache.hc.core5.http.HttpStatus;
-import org.apache.hc.core5.http.ParseException;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 import se.michaelthelin.spotify.SpotifyApi;
 import se.michaelthelin.spotify.enums.ModelObjectType;
-import se.michaelthelin.spotify.exceptions.SpotifyWebApiException;
 import se.michaelthelin.spotify.model_objects.IPlaylistItem;
 import se.michaelthelin.spotify.model_objects.specification.*;
-import se.michaelthelin.spotify.model_objects.credentials.AuthorizationCodeCredentials;
 
-import java.io.IOException;
-import java.time.Instant;
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
 import java.util.*;
@@ -38,92 +28,30 @@ import java.util.stream.Collectors;
 public class SpotifyServiceImpl implements SpotifyService {
     private final SpotifyApi spotifyApi;
     private final SpotifyArtistDao artistDao;
-    private final UserDao userDao;
-
-    @Override
-    public int setAccessToken(String code, HttpServletResponse response) {
-        try {
-            var execute = spotifyApi.authorizationCode(code)
-                    .build()
-                    .execute();
-
-            String accessToken = execute.getAccessToken();
-            String refreshToken = execute.getRefreshToken();
-            long accessTokenExpirationTime = Instant.now().getEpochSecond() + execute.getExpiresIn();
-            spotifyApi.setAccessToken(accessToken);
-
-            String userId = spotifyApi.getCurrentUsersProfile().build().execute().getId();
-            User user = new User(userId, accessToken, refreshToken, accessTokenExpirationTime);
-            userDao.save(user);
-
-            Cookie cookie = new Cookie("user_id", userId);
-            cookie.setPath("/"); // Параметр "Path" вказує шлях, на якому доступне кукі (у цьому випадку доступне на всій домені)
-            response.addCookie(cookie);
-
-            return HttpStatus.SC_OK;
-        } catch (IOException | SpotifyWebApiException | ParseException e) {
-            return HttpStatus.SC_UNAUTHORIZED;
-        }
-    }
+    private final AuthService authService;
 
     @Override
     @SneakyThrows
     @Cacheable(value = CacheConstants.REQUEST_CACHE)
     public PlaylistSimplified[] getPlaylists(HttpServletRequest request) {
-        String accessToken = getAccessTokenFromCookie(request)
-                .orElseThrow(UserUnauthorizedException::new);
-
-        spotifyApi.setAccessToken(accessToken);
-
+        authService.setAccessToken(request);
         return spotifyApi.getListOfCurrentUsersPlaylists()
                 .build()
                 .execute()
                 .getItems();
     }
 
-    private Optional<String> getAccessTokenFromCookie(HttpServletRequest request) {
-        return Optional.ofNullable(request)
-                .map(HttpServletRequest::getCookies)
-                .flatMap(cookies -> Arrays.stream(cookies)
-                        .filter(cookie -> "user_id".equals(cookie.getName()))
-                        .findFirst()
-                        .map(Cookie::getValue))
-                .flatMap(userDao::findById)
-                .map(this::getRefreshedAccessToken);
-    }
-
-    private String getRefreshedAccessToken(User u) {
-        if (Instant.now().getEpochSecond() < u.getAccessTokenExpirationTime()) {
-            return u.getAccessToken();
-        } else {
-            try {
-                AuthorizationCodeCredentials execute = spotifyApi.authorizationCodeRefresh()
-                        .refresh_token(u.getRefreshToken())
-                        .build()
-                        .execute();
-
-                String newAccessToken = execute.getAccessToken();
-                long newAccessTokenExpirationTime = Instant.now().getEpochSecond() + execute.getExpiresIn();
-                u.setAccessToken(newAccessToken);
-                u.setAccessTokenExpirationTime(newAccessTokenExpirationTime);
-                userDao.save(u);
-
-                return newAccessToken;
-            } catch (IOException | ParseException | SpotifyWebApiException e) {
-                return null;
-            }
-        }
-    }
-
-    public List<PlaylistTrack> getRussianTracks(String playlistId) {
+    public List<PlaylistTrack> getRussianTracks(String playlistId, HttpServletRequest request) {
+        authService.setAccessToken(request);
         var russianArtistsFromPlaylist = getRussianArtistsFromPlaylist(playlistId);
         return Arrays.stream(getPlaylistTracks(playlistId))
                 .filter(track -> trackIncludesRussianArtists(track.getTrack().getId(), russianArtistsFromPlaylist))
                 .toList();
     }
 
-    public void removeAllRussianTracksFromPlaylist(String playlistId) {
-        var json = getRussianTracks(playlistId).stream()
+    public void removeAllRussianTracksFromPlaylist(String playlistId, HttpServletRequest request) {
+        authService.setAccessToken(request);
+        var json = getRussianTracks(playlistId, request).stream()
                 .map(PlaylistTrack::getTrack)
                 .map(this::mapToJsonObject)
                 .collect(Collector.of(
@@ -136,7 +64,8 @@ public class SpotifyServiceImpl implements SpotifyService {
         removeFromPlaylist(playlistId, json);
     }
 
-    public void removeTrackFromPlaylist(String playlistId, String trackId) {
+    public void removeTrackFromPlaylist(HttpServletRequest request, String playlistId, String trackId) {
+        authService.setAccessToken(request);
         var json = Arrays.stream(getPlaylistTracks(playlistId))
                 .map(PlaylistTrack::getTrack)
                 .filter(x -> x.getId().equals(trackId))
@@ -216,7 +145,8 @@ public class SpotifyServiceImpl implements SpotifyService {
     }
 
     @Cacheable(value = CacheConstants.REQUEST_CACHE)
-    public Map<String, List<AlbumSimplified>> getLastReleasesFromSubscribedArtists() {
+    public Map<String, List<AlbumSimplified>> getLastReleasesFromSubscribedArtists(HttpServletRequest request) {
+        authService.setAccessToken(request);
         return getAllNewReleases(getUserFollowedArtists());
     }
